extra/mgmt/requirements.txt
../_pmc
../_dnfs
flask
PyJWT
waitress
